// packages/sitemap/src/types.ts
export type SitemapEntry = {
  url: string;
  lastModified?: string | Date;
  changeFrequency?:
    | "always"
    | "hourly"
    | "daily"
    | "weekly"
    | "monthly"
    | "yearly"
    | "never";
  priority?: number;
};

export type RouteGenerator = () => Promise<SitemapEntry[]> | SitemapEntry[];